+++ src/TesterInitialization.php
@@ -1,35 +1,18 @@
 <?php
 
-namespace DeKey\Tester;
+namespace PHPKitchen\CodeSpecs;
 
-use DeKey\Tester\Specification\Tester;
+use PHPKitchen\CodeSpecs\Mixin\TesterInitialization as MainTesterInitialization;
 
 /**
  * Adds support of {@link Tester} in PHPUnit test classes.
  *
- * @package DeKey\Tester
+ * @deprecated use {@link PHPKitchen\CodeSpecs\Mixin\TesterInitialization} instead.
+ * This version kept for BC and would be removed in next major release.
+ *
+ * @package PHPKitchen\CodeSpecs
  * @author Dmitry Kolodko <prowwid@gmail.com>
  */
-trait TesterInitialization {
-    /**
-     * @var Tester instance of tester. This property refreshing before each test.
-     */
-    public $tester;
-
-    /**
-     * @before
-     */
-    public function initTester() {
-        $this->tester = $this->createTester();
-    }
-
-    /**
-     * Creates tester instance before each test.
-     * Override to specify custom tester or do other stuff you want.
-     *
-     * @return Tester tester instance.
-     */
-    public function createTester() {
-        return new Tester($this);
-    }
+trait TesterInitialization  {
+    use MainTesterInitialization;
 }