+++ tests/Base/TestCase.php
@@ -1,8 +1,6 @@
 <?php
 
-namespace Tests\DeKey\Tester\Base;
-
-use DeKey\Tester\TesterInitialization;
+namespace Tests\Base;
 
 /**
  * Represents base class for all of the test cases.
@@ -10,6 +8,6 @@ use DeKey\Tester\TesterInitialization;
  * @author Dmitry Kolodko <prowwid@gmail.com>
  */
 class TestCase extends \PHPUnit\Framework\TestCase {
-    use TesterInitialization;
+    use PHPKitchen\CodeSpecs\Mixin\TesterInitialization;
     const FIXTURES_DIR = __DIR__ . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . 'Fixtures' . DIRECTORY_SEPARATOR;
 }