+++ tests/bootstrap.php
@@ -1,3 +1,4 @@
 <?php
-require_once(dirname(__DIR__) . '/vendor/autoload.php');
+require(dirname(__DIR__) . '/vendor/autoload.php');
+
 error_reporting(E_ALL);